perbaikan migrate update tabel proyek desa

+++ database/migrations/2017_08_23_110353_adding_foreign_tags.php
@@ -13,18 +13,20 @@ class AddingForeignTags extends Migration
      */
     public function up()
     {
-        Schema::table('storytags', function (Blueprint $table) {
-            $table->foreign('story_id') // foreignKey 
-                ->references('id') // dari kolom id 
+        if(!Schema::hasTable('storytags')){
+            Schema::table('storytags', function (Blueprint $table) {
+                $table->foreign('story_id') // foreignKey
+                ->references('id') // dari kolom id
                 ->on('stories') // di tabel users
                 ->onUpdate('cascade') // ketika terjadi perubahan di tabel users maka akan update
                 ->onDelete('cascade'); // ketika data users di hapus akan ikut hilang
-            $table->foreign('tag_id') // foreignKey 
-                ->references('id') // dari kolom id 
+                $table->foreign('tag_id') // foreignKey
+                ->references('id') // dari kolom id
                 ->on('tags') // di tabel users
                 ->onUpdate('cascade') // ketika terjadi perubahan di tabel users maka akan update
                 ->onDelete('cascade'); // ketika data users di hapus akan ikut hilang
-        });
+            });
+        }
     }
 
     /**
